requirements.txt
Flask==3.0.2
gunicorn==20.1.0
python-dotenv==1.0.0
click==8.1.3
Flask==2.1.1
gunicorn==20.1.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
uvicorn==0.22.0
waitress==2.1.2
Werkzeug==2.3.4
discord.py==2.3.2
requests==2.31.0
psycopg=3.1.18
discord-py-interactions:5.11.0